[styeld] 아트박스 상세페에지, 상세모달페이지, 이미지 슬라이드 관련 리펙토링
