+++ src/features/artgram/ArtgramDetail.jsx
@@ -5,55 +5,10 @@ import * as Modal from "./ArtgramModal";
 import { Flex } from "../../components/Flex";
 import { Input } from "../../components/Input";
 import * as Artgramparts from "./Artgramparts";
-// import {BiDotsHorizontalRounded} from 'react-icons/bi'
 import { useGetartgramComments } from "../../hooks/artgram/useGetartgramComments";
 import { decodetoken } from "../../shared/cookies";
 import { usePostingtime } from "../../hooks/artgram/usePostingtime";
-
-// 이미지슬라이더를 위한 
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css"; 
-import "slick-carousel/slick/slick-theme.css";
-import { FaChevronLeft, FaChevronRight } from 'react-icons/fa';
-
-function SamplePrevArrow(props) {
-  const { className, style, onClick } = props;
-  return (
-    <div
-      className={className}
-      style={{ ...style, display: "none" }}
-      onClick={onClick}
-    />
-  );
-}
-
-
-function SampleNextArrow(props) {
-  const { className, style, onClick } = props;
-  return (
-    <div
-      className={className}
-      style={{
-        ...style,
-        display: "block flex",
-        background: "#fffffff6",
-        position: "absolute",
-        right: "10px",
-        width: "50px",
-        height: "50px",
-        justifyContent: "center",
-        alignItems: "center",
-        borderRadius: "50px",
-        color: "black",
-      }}
-      onClick={onClick}
-    >
-      <div style={{position:"relative", right:"8px"}}>
-        <FaChevronRight size={24} color="#e20303" />
-      </div>
-    </div>
-  );
-}
+import ArtgramSlider from "./ArtgramSlider";
 
 function ArtgramDetail({pos}) {
   const { allArtgram, modalState, setModalState } = pos;
@@ -65,22 +20,9 @@ function ArtgramDetail({pos}) {
   );
   const [timehandle] = usePostingtime()  
 
-
-
-  const settings = {
-    speed: 500, // 속도조절
-    // infinite:false,
-    slidesToShow: 1, // 화면에 보여지는 슬라이더의 수 
-    slidesToScroll: 1, // 한번에 넘길 슬라이더이더의 수 
-    style:{width:"100%", position:"relative"},
-    prevArrow: <SamplePrevArrow />,
-    nextArrow: <SampleNextArrow />,
-  };
-
   if (isLoading || isError) {
     return <div>로딩 중....</div>;
   }
-  console.log(data);
   
   return (
     <>
@@ -91,12 +33,7 @@ function ArtgramDetail({pos}) {
         <Flex>
           {/* 상세모달페이지 (1) 이미지 삽입공간 --------------------------------------------------------- */}
           <Modal.ModalinnerImgDiv width="70%">
-            {/* <Modal.ModdalinnerImg src={allArtgram.ArtgramImgs[0].imgUrl} /> */}
-            <Slider {...settings}>
-              {allArtgram.ArtgramImgs.map(img => (
-                <Modal.ModdalinnerImg key={img.imgUrl} src={img.imgUrl} width="100"/>
-              ))}
-            </Slider>
+            <ArtgramSlider map={allArtgram.ArtgramImgs}/>
 
           </Modal.ModalinnerImgDiv>
           {/* 상세모달페이지 (2) 텍스트 공간 ------------------------------------------------------------ */}
